requirements.txt
fastapi==0.116.1
uvicorn[standard]
psycopg2-binary
python-dotenv
sqlalchemy
pandas
pydantic  # remove in case of more basic schema
